% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexfetch.R
\name{hexfetch}
\alias{hexfetch}
\title{Extract Package hexes to target directory}
\usage{
hexfetch(x, dir = NULL)
}
\arguments{
\item{x}{character, name of packages}

\item{dir}{character, name of folder where hexes will be copied to. Will be
created if not available.}
}
\value{
The new path (invisibly).
}
\description{
Extract Package hexes to target directory
}
\examples{
my_dir <- tempdir()
hexfetch("rlang", my_dir)
fs::dir_info(my_dir)
}
